refactor(db-paths): lift DB-path resolvers out of cli into db_paths leaf module (#499 Cluster C)
CLI was the historical home for `db_path`, `_open_store`, and
`_git_common_dir`, so every feature module that needs to open the
brain-graph DB had a function-local `from aelfrice.cli import db_path`.
Static analysis treats those as full edges, closing 10 cycles between
cli and {context_rebuilder, hook_tail, project_warm, mcp_server,
telemetry, hook, doctor, hook_search_tool}.

This commit creates `aelfrice.db_paths` (imports only `aelfrice.store`
+ stdlib) and moves the three resolvers + `DEFAULT_DB_DIR` /
`DEFAULT_DB_FILENAME` / `_ensure_parent_dir` into it. cli.py
re-imports them so `aelfrice.cli.db_path` etc. remain valid for
existing tests and external callers. The 9 feature modules that
previously imported from cli now import from db_paths directly.

After this commit the static cycle-finder reports 0 module-import
cycles ≤ length 4 (down from 14 on main HEAD `25134b7`). The CodeQL
`note` count for module-import-cycle should drop to baseline / zero
on the next PR touching these modules.

No behavior change. 2923 tests pass.
